+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## 2.1.0 (tbd)
 
+- Fixed issue in `parcel-codegen-loader`
 - Added `bun-codegen-loader` package
+- Added restriction in `parcel-plugin-codegen` to only consider `cjs`
 
 ## 2.0.0 (December 26, 2024)